New code to create and update Resume DB
Code contains 2 functions:
1. uploadResume()
This function is used in Upload Resume page. Objective is to 
2. insertResume()
This function is used in both Create Resume and Upload Resume page. Objective is to (a) extract the raw text in the specified document and (b) insert the name, document filename and raw text as a new resume record (row) into resume_db.csv.